requirements.txt
aiohttp
dnspython
python-whois
